feat: use an id as key
This commit is adding a new ID generated when loading the worksheet
instead of using the first columns of the spreadsheet as key.

This will avoid issues when there is several rows with identical values
in the first columns.

The default route is now returning the full worksheet rows instead of
returning only the keys.

This is closer from API standard behavior.

Some search and filter features will have to be added in later commits
to ensure a good user experience.